chore: moved cardinal logic into seperate files
